+++ apps/dashboard/src/app/team/[team_slug]/(team)/~/settings/billing/page.tsx
@@ -1,9 +1,10 @@
+import { ChakraProviderSetup } from "@/components/ChakraProviderSetup";
+import { SettingsBillingPage } from "./BillingSettingsPage";
+
 export default function Page() {
   return (
-    <div className="h-full py-6 container flex items-center justify-center">
-      <h1 className="text-4xl tracking-tighter text-muted-foreground">
-        Billing Settings
-      </h1>
-    </div>
+    <ChakraProviderSetup>
+      <SettingsBillingPage />
+    </ChakraProviderSetup>
   );
 }
