+++ apps/dashboard/src/components/settings/Account/Billing/index.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import {
   type Account,
   AccountPlan,
