+++ apps/dashboard/src/pages/dashboard/settings/billing.tsx
@@ -1,46 +1,16 @@
-import { AccountStatus, useAccount } from "@3rdweb-sdk/react/hooks/useApi";
 import { AppLayout } from "components/app-layouts/app";
-import { Billing } from "components/settings/Account/Billing";
 import { SettingsSidebar } from "core-ui/sidebar/settings";
-import { useRouter } from "next/router";
 import { PageId } from "page-id";
-import { useEffect } from "react";
 import type { ThirdwebNextPage } from "utils/types";
+import { SettingsBillingPage } from "../../../app/team/[team_slug]/(team)/~/settings/billing/BillingSettingsPage";
 
-const SettingsBillingPage: ThirdwebNextPage = () => {
-  const meQuery = useAccount({
-    refetchInterval: (account) =>
-      [
-        AccountStatus.InvalidPayment,
-        AccountStatus.InvalidPaymentMethod,
-      ].includes(account?.status as AccountStatus)
-        ? 1000
-        : false,
-  });
-
-  const router = useRouter();
-  const { data: account } = meQuery;
-
-  const { payment_intent, source_redirect_slug } = router.query;
-
-  // legit usecase, however will move to server side with RSCs
-  // eslint-disable-next-line no-restricted-syntax
-  useEffect(() => {
-    if (payment_intent || source_redirect_slug) {
-      router.replace("/dashboard/settings/billing");
-    }
-  }, [payment_intent, router, source_redirect_slug]);
-
-  if (!account) {
-    return null;
-  }
-
-  return <Billing account={account} />;
+const Page: ThirdwebNextPage = () => {
+  return <SettingsBillingPage />;
 };
 
-SettingsBillingPage.pageId = PageId.SettingsUsage;
+Page.pageId = PageId.SettingsUsage;
 
-SettingsBillingPage.getLayout = (page, props) => (
+Page.getLayout = (page, props) => (
   <AppLayout {...props} hasSidebar={true}>
     <SettingsSidebar activePage="billing" />
 
@@ -48,4 +18,4 @@ SettingsBillingPage.getLayout = (page, props) => (
   </AppLayout>
 );
 
-export default SettingsBillingPage;
+export default Page;
